requirements.txt
altgraph==0.17
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
Pillow==8.0.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
tifffile==2020.10.1
